Bump versions to avoid confusion around releases.
`cargo vendor` has issues with the current type_hash dependencies.

The git repository for v0.3.0 contains a version of type_hash_core
labelled v0.2 but slightly different in repository labels than the v0.2
checked into crates.io. (missing repository, for instance).

`cargo vendor` is pulling v0.2 from crates.io rather than using the
`path = "../core"` version specified in type_hash which means it is
also dropping the repository label on vendored crates but not when the
crate is unvendored.

I'm unclear on what should be the correct semantics here.
This change just bumps all versions to circumvent any issues. It'll
require a new release to be pushed to crates.io as well.